% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phredtonumeric.R
\name{basic_stat}
\alias{basic_stat}
\title{create a data frame of mean, median and quatiles of the quality score per position
Write the basic statistics data frame to csv(current version)
Write the basic statistics data frame to a table in SQLite}
\usage{
basic_stat(name)
}
\arguments{
\item{name}{the object that is the path to the gzipped FASTQ file}
}
\value{
basic statistics data frame

csv file and a SQLite table named
}
\description{
create a data frame of mean, median and quatiles of the quality score per position
Write the basic statistics data frame to csv(current version)
Write the basic statistics data frame to a table in SQLite
}
