% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_over_rep_seq}
\alias{calc_over_rep_seq}
\title{calculate Over Rep seqs}
\usage{
calc_over_rep_seq(infile, out_prefix, min_size = 5L, buffer_size = 1000000L)
}
\arguments{
\item{infile}{A string giving the path for the fastqfile}

\item{out_prefix}{A string giving the prefix to be used for outputs}

\item{min_size}{An int for thhresholding over representation}

\item{buffer_size}{An int for the number of lines to keep in memory}
}
\description{
Description
}
