% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.content.R
\name{contentcycle}
\alias{contentcycle}
\title{Read results nucletoide sequence content matrix from gzipped FASTQ file and save the read matrix to SQLite database}
\usage{
contentcycle()
}
\arguments{
\item{name}{the object that is the path to gzipped file}
}
\value{
data frame obtained by SQLite command from the saved table
}
\description{
Read results nucletoide sequence content matrix from gzipped FASTQ file and save the read matrix to SQLite database
}
