% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.R
\name{convert_long_to_utm}
\alias{convert_long_to_utm}
\title{Return UTM Zone}
\usage{
convert_long_to_utm(lon)
}
\arguments{
\item{lon}{\itemize{
\item Decimal degree longitude value
}}
}
\value{
The function returns a numeric UTM zone (between 1 and 60).
}
\description{
\code{convert_long_2_utm()} take a longitude coordinate and returns the corresponding UTM zone.
}
\details{
Input a longitude (decimal degree) coordinate and this simple function returns the number of the UTM zone where that point falls.
}
