% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCColCheck.R
\name{DCcheck}
\alias{DCcheck}
\title{TDWG Darwin Core Column Name Check}
\usage{
DCcheck(workingdf)
}
\arguments{
\item{workingdf}{- This is dataframe you want to run against the function. To call, simply write workingdf = "the name of your dataframe".}
}
\value{
- what will the function return? A dataframe, string, or something else? The function returns a list of the column names you should fix (because they do not match any of the rules we have established). In addition, a small summary table is printed with the counts of the columns falling under each category (DarwinCore, Custom, DQ, FixMe)
}
\description{
DCcheck checks to see if the column names in your dataframe match the standardized Darwin Core names established by the Taxonomic Databases WorkingGroup
}
\details{
more details on how DC check works/what it does A dataframe is created with all the standard DarwinCore names. The function runs through each of the column names in your working dataframe to see if they match either 1. A standard DarwinCore name 2. A name with a "custom_" indicating a custom made column or 3. A name with a "_DQ" indicating a data quality flag. The function then counts all of the names fitting within each category and creates a summary table.
}
\examples{
# "DCcheck(df)"

}
