% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensions.R
\name{dimensions}
\alias{dimensions}
\title{Extract the dimensions for Fastq file}
\usage{
dimensions(fseq, selection)
}
\arguments{
\item{fseq}{an object that is the read result of the seq.read function}

\item{selection}{"reads' for number of reads/rows, 'positions' for number of positions/columns}
}
\value{
a numeric value of the number of reads or the number of positions
}
\description{
\code{ncolumn}use seqTool to extract the dimensions of a Fastq G zipped file
}
