% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_blanks.R
\name{document_missing_values}
\alias{document_missing_values}
\title{Handles multiple missing values}
\usage{
document_missing_values(
  file_name,
  directory = here::here(),
  colname = NA,
  missing_val_codes = NA,
  replace_value = NA
)
}
\arguments{
\item{file_name}{String. The name of the file to inspect}

\item{directory}{String. Location of file to read/write. Defaults to the current working directory.}

\item{colname}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} String. The columns to inspect. CURRENTLY ONLY WORKS AS SET TO DEFAULT "NA".}

\item{missing_val_codes}{List. A list of strings containing the missing value code or codes to search for.}

\item{replace_value}{String. The value (singular) to replace multiple missing values with. Defaults to NA.}
}
\value{
writes a new dataframe to file. Return invisible.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#questioning}{\figure{lifecycle-questioning.svg}{options: alt='[Questioning]'}}}{\strong{[Questioning]}}
Given a file name (.csv only) and path, the function will search the
columns for any that contain multiple user-specified missing value codes.
For any column with multiple missing value codes, all the missing values
(including blanks) will be replaced with NA. A new column will be generated
and, populated with the given missing value code from the origin column.
Values that were not missing will be populated with "not_missing". The
newly generate column of categorical variables can be used do describe
the various/multiple reasons for why data is absent in the original column.

The function will then write the new dataframe to a file, overwriting the
original file. If it is important to keep a copy of the original file, make
a copy prior to running the function.
}
\details{
Blank cells will be treated as NA.
}
\examples{
\dontrun{
document_missing_values(file_name = "mydata.csv",
                        directory = here::here(),
                        colname = NA, #do not change during function development
                        missing_val_codes = c("missing", "blank", "no data"),
                        replace_value = NA)
                        }
}
