% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.R
\name{.get_unit_boundary}
\alias{.get_unit_boundary}
\title{Gets NPS unit boundaries from Arc GIS}
\usage{
.get_unit_boundary(park_units)
}
\arguments{
\item{park_units}{String. Or list of strings.}
}
\value{
sf dataframe
}
\description{
Gets NPS unit boundaries from Arc GIS
}
\examples{
\dontrun{
.get_unit_boundary("ROMO")
}
}
