% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geographicFuzzing.R
\name{dp_fuzzLocation}
\alias{dp_fuzzLocation}
\title{Convert coordinates for a location into a polygon to obscure specific location}
\usage{
dp_fuzzLocation(lat, lon, coordrefsys, fuzzlevel)
}
\description{
\code{dp_fuzzLocation} "fuzzes" a specific location to something less prciese prior to public release of information
about sensitive resources for which data are not to be released to the public. This function takes coordinates in either UTM
or decimal degrees, converts to UTM (if in decimal degrees), creates a bounding box based on rounding of UTM coordinates to
the 100s or 1000s place, and then creating a polygon from the resultant points. The function returns a string in Well-Known-Text
format.
}
\examples{
dp_fuzzLocation(703977,4035059,32616,"Fuzzed - 1km")
dp_fuzzLocation(36.43909,-84.72429,4326,"Fuzzed - 1km")
}
