% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GC_content.R
\name{GC_content}
\alias{GC_content}
\title{Extract GC content separately and calculate GC content percentage for each sequence read}
\usage{
GC_content(infile)
}
\arguments{
\item{infile}{the object that is the path to the FASTQ file}
}
\value{
plot of GC content
}
\description{
Extract GC content separately and calculate GC content percentage for each sequence read
}
