% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_qc_flags.R
\name{get_custom_flags}
\alias{get_custom_flags}
\title{Create a table of Data Quality Flags based on a custom set of user defined columns headings.}
\usage{
get_custom_flags(directory = here::here(), cols = c(""), force = FALSE)
}
\arguments{
\item{directory}{is the path to the data package .csv files (defaults to the current working directory).}

\item{cols}{A comma delimited list of column names.}

\item{force}{is a logical. Defaults to `FALSE`. When `force = FALSE` the function prints the resulting dataframe to the screen. setting `force = TRUE` suppresses output to the screen.}
}
\value{
a dataframe named cust_flag that contains a row for each column indicated in each .csv file in the directory with the file name, the count of each flag and total number of data points in each .csv (including data flagging columns) and RRU.
}
\description{
get_custom_flags returns a data frame that, for each data file in a data package lists the name of each column selected by the user. By default, any flagged columns are automatically included. The number of each flag type for each column (A, AE, R, P) are reported. Unflagged columns are assumed to have only accepted data. The total number of data points in the specified columns (and data flagging columns for) each .csv are also reported. NAs are excluded. An Unweighted Relative Response (RRU) is calculated as the total number of accepted data points (A, AE, and data that are not flagged) devided by the total number of data points in all specified columns (and the flagged columns).
}
\details{
Flagged columns must have names ending in "*_flag". Missing values must be specified as NA.
}
\examples{
\dontrun{
# specify path and columns by hand:
get_custom_flags("~/my_data_package_directory", cols = c(
  "scientificName",
  "locality"
))

# get custom column names from your csv, use default working directory. Note that in this example column names come from a single .csv but all the .csvs in the directory will be checked for these column names:
cols <- colnames(read.csv("mydata.csv"))[c(1:4, 7, 10)]
get_custom_flags(cols = cols) # if your current working directory IS the data package directory.
}
}
