% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_qc_flags.R
\name{get_dc_flags}
\alias{get_dc_flags}
\title{Create Table of Data Quality Flags in Flagging Columns within individual data columns}
\usage{
get_dc_flags(directory = here::here(), force = FALSE)
}
\arguments{
\item{directory}{is the path to the data package .csv files (defaults to the current working directory).}

\item{force}{is a logical. Defaults to `FALSE`. When `force = FALSE` the function prints the resulting dataframe to the screen. setting `force = TRUE` suppresses output to the screen.}
}
\value{
a dataframe named dc_flag that contains a row for each .csv file in the directory with the file name, the count of each flag and total number of data points in each .csv (including data flagging columns).
}
\description{
get_dc_flags (dc=data columns) returns a data frame that, for each data file in a data package lists the name of each data flagging column and the number of each flag type within that column (A, AE, R, P) as well as the total number of data points in the data flagging columns for each .csv.
}
\details{
The function can be run from within the working directory where the data package is, or the directory can be specified. The function only supports .csv files and assumes that all data flagging columns have column names ending in "_flag". It assumes that there are no additional characters (other than A, AE, R, P) in the flagging cells (such as leading or trailing white spaces). Discrepancies between the sum of the data flagging values and the total number of the cells in the columns may be due to NAs.
}
\examples{
\dontrun{
get_df_flags("~/my_data_package_directory")
get_df_flags() #if your current working directory IS the data package directory.
}

}
