% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_qc_flags.R
\name{get_dp_flags}
\alias{get_dp_flags}
\title{Create Table of Data Quality Flags Found in a Data Package}
\usage{
get_dp_flags(directory = here::here())
}
\arguments{
\item{directory}{is the path to the data package .csv files (defaults to the
current working directory).}
}
\value{
a dataframe named dp_flag that contains the four flags, the count of
each flag and total number of data points in the entire data package.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
get_dp_flags (dp=data package) returns a data frame that list
the number of cells in the entire data package with relevant flags (A, AE,
R, P) as well as the total number of non-NA cells in the data package
(including data flagging columns). Unweighted Relative Response (RRU) is
calculated as the total number of accepted data points (A, AE, and data that
are not flagged).
}
\details{
The function can be run from within the working directory where the
data package is, or the directory can be specified. The function only
supports .csv files and assumes that all .csv files in the folder are part
of the data package. The function counts cells within "*_flag"
columns that start with one of the flagging characters (A, AE, R, P) and
ignores trailing characters and whitespaces. NAs are assumed to be empty
cells or missing data.
}
\examples{
\dontrun{
get_dp_flags("~/my_data_package_directory")
get_dp_flags() # if your current working directory IS the data package
directory.
# ->
get_custom_flags(output="package")
}

}
\keyword{internal}
