% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.R
\name{get_park_polygon}
\alias{get_park_polygon}
\title{Retrieve the polygon information for the park unit from NPS REST services}
\usage{
get_park_polygon(unit_code)
}
\arguments{
\item{unit_code}{is the four-character unit code as designated by NPS.}
}
\description{
\code{get_park_polygon()} retrieves a geoJSON string for a polygon of
a park unit. This is not the official boundary. Note that the REST API call returns the default "convexHull". This is will work better or worse for some parks, depending on the park shape/geography/number of disjunct areas.
#'
}
\examples{
\dontrun{
qc_getParkPolygon("OBRI")
}
}
