% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_user_email}
\alias{get_user_email}
\title{Retrieves an NPS user's email address}
\usage{
get_user_email()
}
\value{
String. The user's email address.
}
\description{
Retrieves an NPS user's email address
}
\details{
The function accesses the system username and then uses a powershell wrapper to access NPS active directory and supply information about the user. That information is then parsed down to the user's email address.

This function probaby won't work for anyone outside of NPS and likely won't work for anyone who is not using a Windows machine. So build those prerequisites (NPS = FALSE) in when calling it from within other function.
}
\examples{
\dontrun{
email <- get_user_email()
}
}
