% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_user_orcid}
\alias{get_user_orcid}
\title{Returns the user's ORCID ID from active directory}
\usage{
get_user_orcid()
}
\value{
Sting. The user's ORCID ID
}
\description{
This is a function to grab a users ORCID from Active Directory. Requires VPN to access AD. If the user does not have an ORCID, returns "NA".
}
\examples{
\dontrun{
orcid <- get_user_orcid()
}
}
