% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.R
\name{get_utm_zone}
\alias{get_utm_zone}
\title{Return UTM Zone}
\usage{
get_utm_zone(lon)
}
\arguments{
\item{lon}{\itemize{
\item Decimal degree longitude value
}}
}
\value{
The function returns a numeric UTM zone (between 1 and 60).
}
\description{
\code{get_utm_zone()} replaces \code{convert_long_2_utm()} as this
function name is more descriptive. \code{get_utm_zone()} takes a longitude
coordinate and returns the corresponding UTM zone.
}
\details{
Input a longitude (decimal degree) coordinate and this simple
function returns the number of the UTM zone where that point falls.
}
