% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer.R
\name{Kmer}
\alias{Kmer}
\title{Return kmer count per sequence for the length of kmer desired}
\usage{
Kmer(path, k)
}
\arguments{
\item{path}{the object that is the path to gzippped FASTQ file}

\item{k}{the length of kmer}
}
\value{
Kmers counts per sequence
}
\description{
Return kmer count per sequence for the length of kmer desired
}
