% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.R
\name{map_WKT}
\alias{map_WKT}
\title{Map WKT geometry (points and polygons)}
\usage{
map_WKT(df, wellknowntext = "footprintWKT", type = "all")
}
\arguments{
\item{df}{\itemize{
\item The name of your data frame contained WKT geometry.
}}

\item{wellknowntext}{\itemize{
\item The name of the column within your data frame containing the WKT geometry.
}}

\item{type}{\itemize{
\item Pick one from "points", "polygons", or "all" to map specific geometry types.
}}
}
\value{
The function returns a dynamic, zoomable leaflet map with the specific geometry plotted.
}
\description{
\code{map_WKT()} takes a well-known text (WKT) geometry column and maps points and polygons onto a gray leaflet map. All NA geometry is dropped before mapping.
}
\details{
Define your dataframe, the column that contains WKT, and an option to map specific geometry types.
}
\examples{
\dontrun{
#map species observations:
map_WKT(my_NPS_species_obs, "footprintWKT")
}

}
