% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncolumn.R
\name{ncolumn}
\alias{ncolumn}
\title{Calculate the number of sequences for Fastq file}
\usage{
ncolumn(name)
}
\arguments{
\item{name}{an object that is the path to the gzippped Fastq file}
}
\value{
a numeric value of the sequence length
}
\description{
\code{ncolumn}use seqTool to extract the sequence length of a Fastq G zipped file
}
