% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrow.R
\name{nreads}
\alias{nreads}
\title{Calculate the number of read for a given Fastq file}
\usage{
nreads(name)
}
\arguments{
\item{name}{an object that is the path to the gzippped Fastq file}
}
\value{
a numeric value of the read count
}
\description{
\code{nreads}use seqTool to extract the number of reads of a Fastq G zipped file
}
