% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overrep_kmer.R
\name{overrep_kmer}
\alias{overrep_kmer}
\title{Generate overrepresented kmers from all kmer counts results}
\usage{
overrep_kmer(path, k, nc, nr)
}
\arguments{
\item{path}{the path to the gz file}

\item{k}{the length of the sequence looking for}

\item{nc}{number of positions}

\item{nr}{number of reads}
}
\value{
the index of reads that has overrepresented kmers
}
\description{
Generate overrepresented kmers from all kmer counts results
}
