% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overrepresented_plot.R
\name{overrep_plot}
\alias{overrep_plot}
\title{Plot the top 5 seqeunces}
\usage{
overrep_plot(overrep_order, prefix)
}
\arguments{
\item{overrep_order}{the table that sorts the sequence content and corresponding counts in descending order}

\item{prefix}{the prefix tot he file saved}
}
\value{
plot of the top 5 overrepresented sequences
}
\description{
Plot the top 5 seqeunces
}
