% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_quality_score_ver2.R
\name{plot_quality_score.2}
\alias{plot_quality_score.2}
\title{Generate a boxplot of the per posistion quality score from basic statistics results}
\usage{
plot_quality_score.2(basic_stat)
}
\arguments{
\item{basic_stat}{the object that is the data frame of the mean, median and quantiles of the FASTQ file from basic statistics function}
}
\value{
boxplot of per position quality score distribution
}
\description{
Generate a boxplot of the per posistion quality score from basic statistics results
}
