% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_All.R
\name{plotall}
\alias{plotall}
\title{Plot all plots about the FASTQ file and save them in a pdf as comprehensive report}
\usage{
plotall(name)
}
\arguments{
\item{name}{the object that is the path to the gzipped FASTQ file}
}
\value{
pdf format report
}
\description{
Plot all plots about the FASTQ file and save them in a pdf as comprehensive report
}
