% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{process_fastq}
\alias{process_fastq}
\title{calculate Over Rep seqs}
\usage{
process_fastq(infile, out_prefix, buffer_size)
}
\arguments{
\item{infile}{A string giving the path for the fastqfile}

\item{out_prefix}{A string giving the prefix to be used for outputs}

\item{buffer_size}{An int for the number of lines to keep in memory}
}
\description{
calculate Over Rep seqs
}
