% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.R
\name{qc_getParkPolygonIRMA}
\alias{qc_getParkPolygonIRMA}
\title{Retrieve the polygon information for the park unit from NPS REST services}
\usage{
qc_getParkPolygonIRMA(Unit_Code)
}
\arguments{
\item{Unit_Code}{is the four-character unit code as designated by NPS.}
}
\description{
\code{qc_getParkPolygon} retrieves a geoJSON string for a polygon of a park unit. This is not the official boundary.
#'
}
\examples{
qc_getParkPolygon("OBRI")
}
