% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.R
\name{qc_ValidateCooridnates}
\alias{qc_ValidateCooridnates}
\title{Check whether a coordinate pair is within the polygon of a park unit}
\usage{
qc_ValidateCooridnates(Unit_Code, lat, lon, validationsource)
}
\arguments{
\item{Unit_Code}{is the four-character unit code as designated by NPS.}

\item{lat}{latitude, in decimal degrees}

\item{lon}{longitude, in decimal degrees.}

\item{validationsource}{is either "IRMA" (for the polygon provided by the NPS rest services), or
"OPBOUNDS" for the park boundary included in the operational boundaries data set (not currently functional).}
}
\description{
\code{qc_ValidateCoordinates} compares a coordinate pair (in decimal degrees) to the polygon for a park unit as provided through the NPS
Units rest services. The function returns a value of TRUE or FALSE.
}
\examples{
qc_ValidateCoordinates("OBRI", 36.07951, -84.65610, "IRMA")
}
