% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{read_adapters}
\alias{read_adapters}
\title{Reads in adapter sequences and returns as an R data frame.}
\usage{
read_adapters(adapters)
}
\arguments{
\item{adapters}{filename of adapter sequences}
}
\value{
A dataframe with columns adapter and sequence
}
\description{
Reads in adapter sequences and returns as an R data frame.
}
