% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeEmptyTables.R
\name{removeEmptyTables}
\alias{removeEmptyTables}
\title{Remove empty tables from a list}
\usage{
removeEmptyTables(df_list)
}
\arguments{
\item{df_list}{A list of tibbles or dataframes.}
}
\value{
The same list but with empty tables removed.
}
\description{
Remove empty tables from a list
}
\examples{

test_list <- list(item_a = tibble::tibble,
                  item_b = mtcars,
                  item_c = iris)

tidy_list <- removeEmptyTables(test_list)

}
