% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.score.R
\name{scorecycle}
\alias{scorecycle}
\title{Read results quality score matrix from gzipped FASTQ file and save the read matrix to SQLite database}
\usage{
scorecycle(name)
}
\arguments{
\item{name}{the object that is the path to gzipped file}
}
\value{
data frame obtained by SQLite command from the saved table
}
\description{
Read results quality score matrix from gzipped FASTQ file and save the read matrix to SQLite database
}
