% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.R
\name{seq_quantile}
\alias{seq_quantile}
\title{\code{seq_quantile} Use seqTools to find the quantile of chosen percentage of quality score per position}
\usage{
seq_quantile(name, q)
}
\arguments{
\item{name}{the object that is the path to the gzipped FASTQ file}

\item{q}{the object that is the percentage of the quanitle}
}
\value{
vector of the per position quantiles of quality score
}
\description{
\code{seq_quantile} Use seqTools to find the quantile of chosen percentage of quality score per position
}
