% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_length.R
\name{sequence_length}
\alias{sequence_length}
\title{extract the sequence length per read and plot corresponding bar plot}
\usage{
sequence_length(fseq)
}
\arguments{
\item{fseq}{the object that is the seqTools processed result}
}
\value{
the plot of the sequence distribution among all reads
}
\description{
extract the sequence length per read and plot corresponding bar plot
}
