% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/te_check.R
\name{te_check}
\alias{te_check}
\title{Threatened Or Endangered Species Checker Function}
\usage{
te_check(x, species_col, park_code, expansion = FALSE)
}
\arguments{
\item{x}{\itemize{
\item The name of your data frame containing species observations
}}

\item{species_col}{\itemize{
\item The name of the column within your data frame containing the scientific names of the species (genus and specific epithet).
}}

\item{park_code}{\itemize{
\item A four letter park code.
}}

\item{expansion}{\itemize{
\item Logical. Defaults to FALSE. The default setting will return only exact matches between your the scientific binomial (genera and specific epithet) in your data set and the federal match list. Setting expansion = TRUE will expand the list of matches to return all species (and subspecies) that from the match list that match any genera listed in your data set, regardless of whether a given species is actually in your data set. An additional column indicating whether the species returned is in your data set ("In Data") or has been expanded to ("Expansion") is generated.
}}
}
\value{
The function returns a (modified) data frame with the names of all the species that fall under the federal conservation list. Technically it is a huxtable, but it should function identically to a data frame for downstream purposes.
}
\description{
\code{te_check()} creates a subset of the species list you provide, all of which fall under the Federal Conservation List Taxa for Data Protection. These species should be removed from your data set if your data is to be made public.
}
\details{
Define your species data set name, column name with the scientific names of your species, and your four letter park code.

The function downloads the Federal Conservation list using the IRMA odata API service and matches this species list to the list of scientific names in your data frame.

Keep in mind that this is a Federal list, not a state list. Changes in taxa names may also cause some species to be missed.

Because the odata API service is not publicly available, you must be logged in to the NPS VPN or in the office to use this function.
}
\examples{
\dontrun{
te_check(df = my_species_dataframe, species_col = "scientificNames", park_code = "BICY")
}

}
