% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TECheck.R
\name{TECheck}
\alias{TECheck}
\title{Threatened Or Endangered Species Checker Function}
\usage{
TECheck(SpeciesData, SciNameCol, ParkID)
}
\arguments{
\item{SpeciesData}{- The name of your dataframe containing species observations}

\item{SciNameCol}{- The name of the column within your dataframe containing the scientific names of the species}

\item{ParkID}{-  The four letter park code}
}
\value{
The function returns a list in your global environment called "TorE" with the names of all the species thnat fall under the federal conservation list.
}
\description{
TECheck creates a subset of the species list you provide, all of which fall under the Federal Conservation List Taxa for Data Protection. These species should be removed from your dataset if your data is to be made public.
}
\details{
Define your species dataset name, column name with the scientific names of your species, and your four letter park code. 

The function downloads the Federal Conservation list using the IRMA odata API service and matches this species list to the list of scientific names in your dataframe.
}
\examples{
TECheck(SpeciesData = my_species_dataframe, SciNameCol = "scientific_names", ParkID = "BICY")
}
