% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.R
\name{test_coord}
\alias{test_coord}
\title{Test whether decimal GPS coordinates are inside a park unit}
\usage{
test_coord(lat, lon, park_unit)
}
\arguments{
\item{lat}{numeric. An individual or vector of numeric values representing the decimal degree latitude of a coordinate}

\item{lon}{numeric. An individual or vector of numeric values representing the decimal degree longitude of a coordinate}

\item{park_unit}{String. Or list of strings each containin the four-letter park unit designation}
}
\value{
dataframe
}
\description{
Test whether decimal GPS coordinates are inside a park unit
}
\examples{
\dontrun{
test_coord(105.555, -47.4332, "DRTO")
}
}
