% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_code_names.R
\name{unit_codes_to_names}
\alias{unit_codes_to_names}
\title{Converts unit codes to full unit (park) names}
\usage{
unit_codes_to_names(unit_code)
}
\arguments{
\item{unit_code}{string or list of strings consisting of (typically) four-letter unit codes.}
}
\value{
a data frame consisting of a single column of full park names
}
\description{
\code{unit_code_to_names} takes a single unit code or vector of unit codes and returns a data frame of full unit names using a public IRMA API. For example if given the code "ROMO" the function will return "Rocky Mountain National Park".
}
\examples{
\dontrun{
 unit_codes_to_names("ROMO")
 unit_codes_to_names(c("ROMO", "GRYN"))
 }
}
