% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.R
\name{validate_coord}
\alias{validate_coord}
\title{Check whether a coordinate pair is within the polygon of a park unit}
\usage{
validate_coord(unit_code, lat, lon)
}
\arguments{
\item{unit_code}{is the four-character unit code as designated by NPS.}

\item{lat}{latitude, in decimal degrees}

\item{lon}{longitude, in decimal degrees.}
}
\description{
\code{validate_coord()} compares a coordinate pair (in decimal degrees) to the polygon for a park unit as provided through the NPS
Units rest services. The function returns a value of TRUE or FALSE.
}
\examples{
\dontrun{
qc_ValidateCoordinates("OBRI", 36.07951, -84.65610)
}
}
