% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.R
\name{validate_coord_list}
\alias{validate_coord_list}
\title{Test whether decimal GPS coordinates are inside a park unit}
\usage{
validate_coord_list(lat, lon, park_units)
}
\arguments{
\item{lat}{numeric. An individual or vector of numeric values representing
the decimal degree latitude of a coordinate}

\item{lon}{numeric. An individual or vector of numeric values representing
the decimal degree longitude of a coordinate}

\item{park_units}{String. Or list of strings each containing the four letter
park unit designation}
}
\value{
logical
}
\description{
This function can take a list of coordinates and park units as input. In
addition to being vectorized, depending on the park borders, it can be a
major improvement on \code{validate_coord()}.
}
\examples{
\dontrun{
x <- validate_coord_list(lat = 105.555, long = -47.4332, park_units = "DRTO")

# or a dataframe with many coordinates and potentially many park units:
x <- validate_coord_list(lat = df$decimalLatitutde,
                lon = df$decimalLongitude,
                park_units = df$park_units)
# you can then merge it back in to the original dataframe:
df$test_GPS_coord <- x
}
}
