% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_stat.R
\name{basic_stat}
\alias{basic_stat}
\title{Generate the data frame that includes percentiles of quality score per position}
\usage{
basic_stat(infile)
}
\arguments{
\item{infile}{the object that is the dataframe of the mean, median and quantiles of the FASTQ file from basic statistics function}
}
\value{
boxplot of per position quality score distribution
}
\description{
Generate the data frame that includes percentiles of quality score per position
}
