% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GC_content_plot.R
\name{GC_content}
\alias{GC_content}
\title{Extract GC content and construct a plot of GC content per position
\code{GC_content} Use SQLite table to extract per position GC content and use ggplot for plotting}
\usage{
GC_content(name)
}
\arguments{
\item{name}{the object that is the path to the gzipped FASTQ file}
}
\value{
plot of GC content
}
\description{
Extract GC content and construct a plot of GC content per position
\code{GC_content} Use SQLite table to extract per position GC content and use ggplot for plotting
}
