% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_cols.R
\name{order_cols}
\alias{order_cols}
\title{Ordering Columns Function 03-21-2023}
\usage{
order_cols(df)
}
\arguments{
\item{df}{\itemize{
\item This is the dataframe you want to run against the function. To call, simply write df = "the name of your dataframe".
}}
}
\value{
\itemize{
\item The function returns a list of required and suggested columns to include in your dataset. When assigning to an object, the object contains your new dataset with all columns ordered properly.
}
}
\description{
\code{order_cols()} Checks and orders columns with TDWG Darwin Core naming standards and custom names in a dataset
}
\details{
Check to see if you have the three required columns (locality, type, basisOfRecord) and various suggested columns present in your dataset. Print a list of which columns are present and which are not. Then, order all the columns in your dataset in the following order: required columns, suggested columns, the rest of the Darwin Core columns, "custom_" (non-Darwin Core) columns, and finally sensitive species data columns.
}
\examples{
\dontrun{
order_cols(df)
}
}
