% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overrep_kmer.R
\name{overrep_kmer}
\alias{overrep_kmer}
\title{Generate overrepresented kmers from all kmer counts results}
\usage{
overrep_kmer(fseq, fseq_count, k, name, nc, nr)
}
\arguments{
\item{fseq}{the read result from seqTools}

\item{fseq_count}{kmer result generated from kmer function}

\item{k}{the length of the sequence looking for}

\item{name}{the path to the data}

\item{nc}{number of positions}

\item{nr}{number of reads}
}
\value{
the index of reads that has overrepresented kmers
}
\description{
Generate overrepresented kmers from all kmer counts results
}
