% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_perseq_quality.R
\name{plot_perseq_quality}
\alias{plot_perseq_quality}
\title{plot the mean quality score per read
\code{plot_perseq_quality} use ggplot to plot the mean quality score per read}
\usage{
plot_perseq_quality(scorecycle)
}
\arguments{
\item{scorecycle}{the object that is the matrix of quality score read from the read score function}
}
\value{
plot of mean quality score per read
}
\description{
plot the mean quality score per read
\code{plot_perseq_quality} use ggplot to plot the mean quality score per read
}
