% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_perseq_quality.R
\name{plot_perseq_quality}
\alias{plot_perseq_quality}
\title{plot the mean quality score per read in histograms
\code{plot_perseq_quality}}
\usage{
plot_perseq_quality(infile)
}
\arguments{
\item{infile}{the object that is the path to the file that}
}
\value{
plot of mean quality score per read
}
\description{
plot the mean quality score per read in histograms
\code{plot_perseq_quality}
}
