% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_quality_score_ver2.R
\name{plot_quality_score.2}
\alias{plot_quality_score.2}
\title{Generate a boxplot of the per posistion quality score from basic statistics results}
\usage{
plot_quality_score.2(name)
}
\arguments{
\item{name}{the object that is the path to the gzipped FASTQ file}
}
\value{
boxplot of per position quality score distribution
}
\description{
Generate a boxplot of the per posistion quality score from basic statistics results
}
