% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_quality_score_ver2.R
\name{plot_quality_score}
\alias{plot_quality_score}
\title{Generate a boxplot of the per posistion quality score from basic statistics results}
\usage{
plot_quality_score(basic_statistics, writefile = FALSE, prefix)
}
\arguments{
\item{basic_statistics}{the object that is the dataframe of the mean, median and quantiles of the FASTQ file from basic statistics function}

\item{writefile}{the object indicating intent to save the plot as pdf file, set default as FALSE}

\item{prefix}{the prefix for the output file of the plot}
}
\value{
boxplot of per position quality score distribution
}
\description{
Generate a boxplot of the per posistion quality score from basic statistics results
}
