% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sequence_content.R
\name{plotSeqContent}
\alias{plotSeqContent}
\title{Plot the per position nucleotide content
\code{plotSeqContent}}
\usage{
plotSeqContent(name)
}
\arguments{
\item{name}{the object that is the path to the gzipped FASTQ file}
}
\value{
ggplot line plot of all nucleotide content inclding  A, T, G, C and N
}
\description{
Plot the per position nucleotide content
\code{plotSeqContent}
}
