% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sequence_content.R
\name{plotSeqContent}
\alias{plotSeqContent}
\title{Plot the per position nucleotide content
\code{plotSeqContent}}
\usage{
plotSeqContent(fseq, nr, nc)
}
\arguments{
\item{nr}{the number of reads of the FASTQ file, acquired through previous functions}

\item{nc}{the number of positions of the FASTQ file, acquired through previous functions}

\item{name}{the object that is the path to the gzipped FASTQ file}
}
\value{
ggplot line plot of all nucleotide content inclding  A, T, G, C and N
}
\description{
Plot the per position nucleotide content
\code{plotSeqContent}
}
