% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCkit-package.R
\docType{package}
\name{QCkit-package}
\alias{QCkit}
\alias{QCkit-package}
\title{QCkit: NPS Inventory and Monitoring Quality Control Toolkit}
\description{
This package contains a set of useful functions for data munging, quality control, and data flagging. Functions are contributed by multiple U.S. National Park Service staff, contractors, partners and others. These functions will likely be most useful for quality control of NPS data but may have utility beyond their intended functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nationalparkservice/QCkit/}
  \item Report bugs at \url{https://nationalparkservice.github.io/QCkit/}
}

}
\author{
\strong{Maintainer}: Robert Baker \email{robert_baker@nps.gov} (\href{https://orcid.org/0000-0001-7591-5035}{ORCID})

Authors:
\itemize{
  \item Judd Patterson \email{judd_patterson@nps.gov} (\href{https://orcid.org/0000-0002-0951-7917}{ORCID})
  \item Joe DeVivo
  \item Issac Quevedo (\href{https://orcid.org/0000-0003-0129-981X}{ORCID})
  \item Sarah Wright (\href{https://orcid.org/0009-0004-5060-2189}{ORCID})
}

Other contributors:
\itemize{
  \item Sarah Kelso (\href{https://orcid.org/0009-0002-8468-6945}{ORCID}) [contributor]
  \item Amy Sherman (\href{https://orcid.org/0000-0002-9276-0087}{ORCID}) [contributor]
}

}
\keyword{internal}
