% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_content.R
\name{sequence_content}
\alias{sequence_content}
\title{Extract nucleoctide sequence content per position from fastq file}
\usage{
sequence_content(fseq, content)
}
\arguments{
\item{fseq}{an object that is the read result from seq.read function}

\item{content}{an object of string type that specifies the content in question, "A","T","G","C","N"(either capital or lower case)}
}
\value{
the per position
}
\description{
Extract nucleoctide sequence content per position from fastq file
}
\author{
Wenyue Xing, \email{wenyue_xing@brown.edu}
}
