% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/te_check.R
\name{te_check}
\alias{te_check}
\title{Threatened Or Endangered Species Checker Function}
\usage{
te_check(df, species_col, park_code)
}
\arguments{
\item{df}{- The name of your dataframe containing species observations}

\item{species_col}{- The name of the column within your dataframe containing the scientific names of the species}

\item{park_code}{-  The four letter park code}
}
\value{
The function returns a list in your global environment called "TorE" with the names of all the species that fall under the federal conservation list.
}
\description{
te_check creates a subset of the species list you provide, all of which fall under the Federal Conservation List Taxa for Data Protection. These species should be removed from your dataset if your data is to be made public.
}
\details{
Define your species dataset name, column name with the scientific names of your species, and your four letter park code.

The function downloads the Federal Conservation list using the IRMA odata API service and matches this species list to the list of scientific names in your dataframe.

Keep in mind that this is a Federal list, not a state list. Changes in taxa names may also cause some species to be missed.

Because the Federal Conservation list is not publicly available (???), you must be logged in to the NPS VPN or in the office to use this function.
}
\examples{
\dontrun{
te_check(df = my_species_dataframe, species_col = "scientificNames", park_code = "BICY")
}

}
